% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_target_score.R
\name{calc_target_score}
\alias{calc_target_score}
\title{Compute Target score for randomized data/compute P value for each TS given a network topology}
\usage{
calc_target_score(wk, wks, dist_ind, inter, n_dose, n_prot,
  proteomic_responses, max_dist = 1, cell_line, verbose = TRUE,
  ts_factor = 1, fs_file, antibody_map_file = NULL, dist_file = NULL)
}
\arguments{
\item{wk}{TODO}

\item{wks}{TODO}

\item{dist_ind}{TODO}

\item{inter}{TODO}

\item{n_dose}{TODO}

\item{n_prot}{TODO}

\item{proteomic_responses}{TODO}

\item{max_dist}{TODO (default: 1)}

\item{cell_line}{TODO}

\item{verbose}{a boolean to show debugging information}

\item{ts_factor}{a scaling factor for the pathway component in the target score}

\item{fs_file}{Functional score file. A tab-delmited file with a header, each row is an
antibody in the first column and functional score in the second column
(i.e. 1 oncogene, 0 tumor supressor/oncogene, -1 tumor supressor characteristics)}

\item{antibody_map_file}{a listing of antibodies, their associated genes, and modification sites}

\item{dist_file}{A distance file an edgelist with a third column which is the network distance
between the genes in the interaction}
}
\description{
Compute Target score for randomized data/compute P value for each TS given a network topology
}
\details{
data: multiple dose single drug perturbation
ts: integral_dose(fs*(xi+sigma_j(2^p*xj*product_k(wk))))
missing: For phosp and dephosp based wk, there is no 'exact match' between known and measured phospho-sites
}
\concept{zeptosensPkg}
