% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_sif_from_matrix.R
\name{create_sif_from_matrix}
\alias{create_sif_from_matrix}
\title{Create sif from network matrix.}
\usage{
create_sif_from_matrix(t_net, genelist)
}
\arguments{
\item{t_net}{p*p Gaussian gene Network matrix estimated.}

\item{genelist}{Gene name list corresponding to the Gaussian gene Network estimated.}
}
\value{
Edgelist description data.frame contain G(V,E) including Vertex and Edges
(1 positive/-1 negative) edgevalue and vertex number of Gaussian Graphical Model.
}
\description{
Create sif from network matrix.
}
\concept{zeptosensPkg}
