% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSifFromMatrix.R
\name{createSifFromMatrix}
\alias{createSifFromMatrix}
\title{Create sif from network matrix.}
\usage{
createSifFromMatrix(tNet, genelist)
}
\arguments{
\item{tNet}{p*p Gaussian gene Network matrix estimated.}

\item{genelist}{gene name list corresponding to the Gaussian gene Network estimated.}
}
\value{
edgelist description data.frame contain G(V,E) including Vertex and Edges 
(1 positive/-1 negative) edgevalue and vertex number of Gaussian Graphical Model.
}
\description{
Create sif from network matrix.
}
\concept{zeptosensPkg}
