% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getChibeExport.R
\name{genChibeExport}
\alias{genChibeExport}
\title{Generate ChiBE Export}
\usage{
genChibeExport(antibodies, antibodyMapFilename = NULL,
  chibeFilename = "abChibe.txt")
}
\arguments{
\item{antibodies}{a vector of strings of antibody names}

\item{chibeFilename}{a string, the name of the file with gene/site information needed for ChiBE}

\item{antibodyMap}{a string, antibody map file name 
(default: NULL, if NULL then use zentosensPkg internal antibody map)}
}
\description{
Generate ChiBE Export
}
\concept{zeptosensPkg}
