% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateCausalityGraph.R
\name{generateCausalityGraph}
\alias{generateCausalityGraph}
\title{Generate Causality Graph}
\usage{
generateCausalityGraph(platformFile, idColumn = "NodeName",
  symbolsColumn = "Symbol", sitesColumn = "Sites",
  effectColumn = "Effect", valuesFile, valueColumn = "Value",
  valueThreshold, graphType = "compatible", doSiteMatch = TRUE,
  siteMatchProximityThreshold = 0, siteEffectProximityThreshold = 0,
  geneCentric = TRUE, colorSaturationValue = 10, outputFilePrefix,
  customNetworkDirectory = tempdir())
}
\arguments{
\item{platformFile}{Name of the antibody reference file}

\item{idColumn}{Column name of IDs}

\item{symbolsColumn}{Column name of gene symbols}

\item{sitesColumn}{Column name for phosphorylation sites}

\item{effectColumn}{Column name for effect of the site on activity}

\item{valuesFile}{Name of the measurements file}

\item{valueColumn}{Name of the values column in the measurements file}

\item{valueThreshold}{The value threshold to be considered as significant}

\item{graphType}{Either 'compatible' or 'conflicting'}

\item{doSiteMatch}{option to enforce matching a phosphorylation site in the network with the annotation of antibody}

\item{siteMatchProximityThreshold}{when site matching is on, this parameter sets the proxomity threshold for a
site number in the relation to match the site whose change is observed in the data (DEFAULT: 0)}

\item{siteEffectProximityThreshold}{when the site effect is not known, we can approximate it with the known
effect of proximate sites. This parameter sets the proximity threshold for using the proximate sites for that 
prediction. (DEFAULT: 0)}

\item{geneCentric}{A boolean option to produce a gene-centric (TRUE) or an antibody-centric (FALSE) graph}

\item{colorSaturationValue}{The value that maps to the most saturated color (DEFAULT: 10)}

\item{outputFilePrefix}{If the user provides xxx, then xxx.sif and xxx.format are generated}

\item{customNetworkDirectory}{The directory that the network will be downloaded and SignedPC
directory will be created in. Pass null to use default.}
}
\description{
Generate Causality Graph
}
\concept{
zeptosensPkg
}
