% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateCausalityIgraph.R
\name{generateCausalityIgraph}
\alias{generateCausalityIgraph}
\title{Generate igraph Graph from Causality Analysis Results}
\usage{
generateCausalityIgraph(causalityResults)
}
\arguments{
\item{causalityResults}{results generated by generateRPPAGraphs()}
}
\value{
igraph graph object
}
\description{
Generate igraph Graph from Causality Analysis Results
}
\concept{zeptosensPkg}
