% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateRPPAGraphs.R
\name{generateRPPAGraphs}
\alias{generateRPPAGraphs}
\title{Generate RPPA Graphs}
\usage{
generateRPPAGraphs(platformFile, idColumn = "NodeName",
  symbolsColumn = "Symbol", sitesColumn = "Sites",
  effectColumn = "Effect", valuesFile, valueColumn = "Value",
  valueThreshold, graphType = "compatible", siteMatchStrict = TRUE,
  geneCentric = TRUE, outputFilePrefix, customNetworkDirectory = tempdir())
}
\arguments{
\item{platformFile}{Name of the antibody reference file}

\item{idColumn}{Column name of IDs}

\item{symbolsColumn}{Column name of gene symbols}

\item{sitesColumn}{Column name for phosphorylation sites}

\item{effectColumn}{Column name for effect of the site on activity}

\item{valuesFile}{Name of the measurements file}

\item{valueColumn}{Name of the values column in the measurements file}

\item{valueThreshold}{The value threshold to be considered as significant}

\item{graphType}{Either 'compatible' or 'conflicting'}

\item{siteMatchStrict}{boolean option to enforce matching a phosphorylation site in the network with
the annotation of antibody}

\item{geneCentric}{A boolean option to produce a gene-centric (TRUE) or an antibody-centric (FALSE) graph}

\item{outputFilePrefix}{If the user provides xxx, then xxx.sif and xxx.format are generated}

\item{customNetworkDirectory}{The directory that the network will be downloaded and SignedPC
directory will be created in. Pass null to use default.}
}
\description{
Generate RPPA Graphs
}
\concept{
zeptosensPkg
}

