% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fs_vals.R
\name{get_fs_vals}
\alias{get_fs_vals}
\title{Extracted functional score value from COMIC/ONCODRIVE Database. Can be override with Manually set functional score.}
\usage{
get_fs_vals(
  n_prot,
  proteomic_responses,
  mab_to_genes,
  fs_override = NULL,
  verbose = FALSE
)
}
\arguments{
\item{n_prot}{Antibody number of input data.}

\item{proteomic_responses}{Input drug perturbation data. With columns as antibody, rows as samples.}

\item{mab_to_genes}{A list of antibodies, their associated genes, modification sites and effect.}

\item{fs_override}{a listing of functional scores for each gene manually set up
for overriding COSMIC Database given value, the modification path. (.txt)}

\item{verbose}{Default as FALSE. If given TRUE, will print out the gene seq mapped with Antibody Map File.}
}
\description{
Extracted functional score value from COMIC/ONCODRIVE Database. Can be override with Manually set functional score.
}
\concept{zeptosensPkg}
