% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_target_score.R
\name{get_target_score}
\alias{get_target_score}
\title{Target Score Pilot Run}
\usage{
get_target_score(wk, wks, dist_ind, inter, n_dose, n_prot,
  proteomic_responses, max_dist = 1, n_perm, cell_line,
  target_score_output_file = NULL, matrix_wk_output_file = NULL,
  target_score_q_value_file = NULL, target_score_dose_file = NULL,
  random_target_score_file = NULL, target_score_p_value_file = NULL,
  verbose = TRUE, ts_factor = 1, fs_file,
  signed_matrix_wk_output_file = NULL)
}
\arguments{
\item{wk}{inference network constructed in matrix form with edge strength value estimated.
Can be extracted directly from predict_bio_network,predict_dat_network or predictt_hyb_network
function. Where predict_bio_network edge value default at 1 for upregulate and -1 for down regulate.}

\item{wks}{inference network constructed in matrix form with edge strength value estimated.
Can be extracted directly from predict_bio_network,predict_dat_network or predictt_hyb_network
function. Where predict_bio_network edge value default at 1 for upregulate, -1 for down regulate,
2 for phosphorylation and -2 for dephosphorylation.}

\item{dist_ind}{i distance file of edgelist with a third column as the network distance
between the genes in the interaction}

\item{inter}{edgelist of inferred network.}

\item{n_dose}{dose number of input data.}

\item{n_prot}{antibody number of input data.}

\item{proteomic_responses}{input drug perturbation data. With columns as antibody, rows as samples.}

\item{max_dist}{maximum edge strength value.(Default at 1)}

\item{n_perm}{number of random TS calculations for building the null distribution}

\item{cell_line}{cell lines labels included in input file}

\item{target_score_output_file}{a filename to write total target score results (default: NULL)}

\item{matrix_wk_output_file}{a filename to write matrix inferred network wk (default: NULL)}

\item{target_score_q_value_file}{a filename to write statistical significance levels (default: NULL)}

\item{target_score_dose_file}{a filename to write dose dependent target score results (default: NULL)}

\item{random_target_score_file}{a filename to write random target score generated by random network (default: NULL)}

\item{target_score_p_value_file}{a filename to write statistical significance levels (default: NULL)}

\item{verbose}{a flag for debugging output}

\item{ts_factor}{a scaling factor for the pathway component of the target score}

\item{fs_file}{a file with the functional score data}

\item{signed_matrix_wk_output_file}{a filename to write matrix inferred network wks (default: NULL)}
}
\description{
Target Score Pilot Run
}
\details{
data: multiple dose single drug perturbation
ts: integral_dose(fs*(xi+sigma_j(2^p*xj*product_k(wk))))
to be converted to integral_dose(fs*((xi/(stdev_xi)+sigma_j(2^p*((xj/stdev_xj)*product_k(wk))))
missing: For phosp and dephosp based wk, there is no 'exact match' between known and measured phospho-sites
}
\concept{zeptosensPkg}
