% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVocanoPot.R
\name{getVocanoPlot}
\alias{getVocanoPlot}
\title{Plot vocano plot of Target Score Result. As Significant Proteins (p value< Defalut or Manually set value) will show in red. The inverse log 10 of Target Score q value and Target Score calculated were shown.}
\usage{
getVocanoPlot(TS, Qvalue, filename, path = NULL)
}
\arguments{
\item{TS}{input Target Score calculated for each protein data frame. Gene in coloumns and samples in row. Wih colnames as gene tags and rownames as sample tags.}

\item{Qvalue}{input Target Score q value calculated for each protein data frame.Gene in coloumns and samples in row. Wih colnames as gene tags and rownames as sample tags.}

\item{filename}{Manually set filename of Vocano Plot.}

\item{path}{Plot Store path. Default at working environment.}
}
\value{
Vocano Plots with indicated filename and path.
}
\description{
Plot vocano plot of Target Score Result. As Significant Proteins (p value< Defalut or Manually set value) will show in red. The inverse log 10 of Target Score q value and Target Score calculated were shown.
}
\examples{
optimizeParameter(data=GeneExpresssion,prior=Priorindormation)
}
\concept{zeptosensPkg}
