% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchGenesToEdgelist.R
\name{matchGenesToEdgelist}
\alias{matchGenesToEdgelist}
\title{Match a vector of genes to a SIF network}
\usage{
matchGenesToEdgelist(genes1, genes2 = NULL, annotEdgelist, antibodyVec,
  useAnnot = FALSE, verbose = FALSE)
}
\arguments{
\item{genes1}{a vector of genes}

\item{genes2}{an optional vector if the source}

\item{annotEdgelist}{a data.frame; the first two columns are interaction 
participants and the third column is an optional annotation}

\item{antibodyVec}{a vector with the names of the antibodies (e.g. colnames(proteomicResponses)); 
indicies returned from this function will be mapped to this vector}

\item{useAnnot}{a boolean as to whether the optional annotation column values 
in the annotEdgelist should outputted in the results}

\item{verbose}{a boolean to show debugging information}
}
\value{
a data.frame, columns 1-2 are indicies of the edgelist participants, 
  column 3 is the annotation values for the edgelist, 4-5 the names of the 
  edgelist participants
}
\description{
Match a vector of genes to a SIF network
}
\concept{zeptosensPkg}
