% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{network}
\alias{network}
\title{Network from signedPC}
\usage{
network(nProt, proteomicResponses, maxDist, antibodyMapFile = NULL,
  distFile = NULL, verbose = F)
}
\arguments{
\item{nProt}{TBA}

\item{proteomicResponses}{TBA}

\item{maxDist}{TBA (default: 1)}

\item{antibodyMapFile}{a listing of antibodies, their associated genes, and modification sites}

\item{distFile}{A distance file an edgelist with a third column which is the network distance
between the genes in the interaction}
}
\description{
Network from signedPC
}
\examples{

}
\concept{zeptosensPkg}
