% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_parameter_dat.R
\name{optimize_parameter_dat}
\alias{optimize_parameter_dat}
\title{Choose the optimal regulization parameter for only data-driven network.}
\usage{
optimize_parameter_dat(data, rho = seq(0.01, 1, length = 100))
}
\arguments{
\item{data}{input proteomics dataset for network inference. Gene in coloumns and samples in row.
With colnames as gene tags and rownames as sample tags.}

\item{rho}{positive tuning parameter for elastic net penalty. Default at seq(0.01,1,length=100).}
}
\value{
Parameter of regulization BIC error calculated.
\item{rho}{penalty parameter.}
\item{bic}{BIC error calculated for penalty parameters.}
}
\description{
Choose the optimal regulization parameter for only data-driven network.
}
\concept{zeptosensPkg}
