% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_parameter_hybrid.R
\name{optimize_parameter_hybrid}
\alias{optimize_parameter_hybrid}
\title{Choose the optimal regulization parameter and scale paramter for adjusted-glasso algorithm network construction}
\usage{
optimize_parameter_hybrid(data, prior, rho = seq(0.01, 1, length = 100),
  kappa = seq(0.01, 1, length = 100))
}
\arguments{
\item{data}{input proteomics dataset for network inference. Gene in coloumns and samples in row.
With colnames as gene tags and rownames as sample tags.}

\item{prior}{Prior information matrix of gene interaction, with colnames and rownames as gene tags.
With colnames and rownames as gene tags.Can be inferred from Public data source (for example:SignedPC).}

\item{rho}{positive tuning parameter for elastic net penalty. Default at seq(0.01,1,length=100).}

\item{kappa}{positive scaler parameter for biology-knowledge base contribution. Default at seq(0.01,1,length=100).}
}
\value{
Parameter list of regulization parameter decided by the prior information and the algorithmn lowest BIC.
\item{rho}{optimized penalty parameter.}
\item{kappa}{optimized scaler parameter.}
\item{rho_m}{optimized regulization matrix for the expression data.}
\item{bic}{Calculated model BIC error for differnet regularization parameters.}
}
\description{
Choose the optimal regulization parameter and scale paramter for adjusted-glasso algorithm network construction
}
\concept{zeptosensPkg}
