% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizeParameter2.R
\name{optimizeParameter2}
\alias{optimizeParameter2}
\title{Choose the optimal regulization parameter for only data-driven network.}
\usage{
optimizeParameter2(data)
}
\arguments{
\item{data}{input expression data frame. Gene in coloumns and samples in row.
With colnames as gene tags and rownames as sample tags.}
}
\value{
Parameter of regulization decided lowest BIC. Including regularize parameter
(L1 norm parameter) as rho.
}
\description{
Choose the optimal regulization parameter for only data-driven network.
}
\examples{
optimizeParameter(data = GeneExpresssion, prior = Priorindormation)

}
\concept{zeptosensPkg}
