% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseCausalityToIgraph.R
\name{parseCausalityToIgraph}
\alias{parseCausalityToIgraph}
\title{Parse Causality Results to igraph}
\usage{
parseCausalityToIgraph(results)
}
\arguments{
\item{results}{generateCausalityGraph() results; list with sif and format data.frames}
}
\value{
an igraph object
}
\description{
Parse Causality Results to igraph
}
\concept{
zeptosensPkg
}
