% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_dat_network.R
\name{predict_dat_network}
\alias{predict_dat_network}
\title{predict data-driven only network.}
\usage{
predict_dat_network(data, cut_off = 0.1, n_prot, proteomic_responses,
  max_dist = 1)
}
\arguments{
\item{data}{input expression data frame. Gene in coloumns and samples in row.
With colnames as gene tags and rownames as sample tags.}

\item{cut_off}{Manually Setup cut off value for the strength of edge. Default at 0.1.}

\item{n_prot}{Number of proteins included in the drug perturbation data.}

\item{proteomic_responses}{Drug Perturbation data for analysis.}

\item{max_dist}{TODO (default: 1)}
}
\value{
Parameter of regulization decided lowest BIC.Including regularize parameter(L1 norm parameter) as rho.
}
\description{
predict data-driven only network.
}
\concept{zeptosensPkg}
