% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictBioNetwork.R
\name{predictBioNetwork}
\alias{predictBioNetwork}
\title{Predict Network from signedPC (Bio-inferred network)}
\usage{
predictBioNetwork(nProt, proteomicResponses, maxDist,
  antibodyMapFile = NULL, distFile = NULL, verbose = F)
}
\arguments{
\item{nProt}{TBA}

\item{proteomicResponses}{TBA}

\item{maxDist}{TBA (default: 1)}

\item{antibodyMapFile}{a listing of antibodies, their associated genes, and modification sites}

\item{distFile}{A distance file an edgelist with a third column which is the network distance
between the genes in the interaction}
}
\description{
Predict Network from signedPC (Bio-inferred network)
}
\examples{

}
\concept{zeptosensPkg}
