% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictDirectionalNetwork.R
\name{predictDirectionalNetwork}
\alias{predictDirectionalNetwork}
\title{Predicted Directional GeneNetwork}
\usage{
predictDirectionalNetwork(data, prior, rho, kappa, cut_off)
}
\arguments{
\item{data}{input expression data. Coloumns as the gene, rows as the sample.
With colnames as the gene tags, rownames as the sample tags.}

\item{prior}{prior information matrix with colnames and rownames as the gene tags.
Can extracted from predictBioNetwork function or inferred from any resources.}

\item{rho}{regulization parameter}

\item{kappa}{scaler parameter}

\item{cut_off}{TODO}
}
\value{
estimated_network include list of estimated directional partial correlation
gene network rho as the regulization parametwe and kappa as the scaler parameter.
}
\description{
Predicted Directional GeneNetwork
}
\concept{zeptosensPkg}
