% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictHybNetwork.R
\name{predictHybNetwork}
\alias{predictHybNetwork}
\title{Choose the optimal regulization parameter and scale paramter for prior information adjusted network construction.}
\usage{
predictHybNetwork(data, prior = NULL, cut_off = 0.1,
  proteomicResponses, nProt, maxDist = 1, antibodyMapFile)
}
\arguments{
\item{data}{input Global Signaling expression data frame(TCGA rppa Data for example).
Gene in coloumns and samples in row. Wih colnames as gene tags and rownames as sample tags.}

\item{prior}{Prior information data frame ,with colnames and rownames as gene tags.
With colnames and rownames as gene tags. Can be inferred from predictBioNetwork or any network resources.}

\item{cut_off}{Manually set up cut off value for strength of edge. Default at 0.1.}

\item{proteomicResponses}{RPPA data tested for drug pertubation.}

\item{nProt}{number of Proteins contained in the data.}

\item{maxDist}{Maximun distance for the network. Default at 1.}

\item{antibodyMapFile}{TBA}
}
\value{
"parameters" as the Parameter list of regulization parameter decided by the prior information and the algorithmn lowest BIC.Including regularize parameter(L1 norm parameter) as "rho", scale parameter(decided how much prior information contribute) as "kappa", and regulization matrix for the expression data as "rhoM".

"bic"as the Model's BIC matrix for differnet regularization parameters.

"wk" as the predicted network.

"wks" TBA

"distInd" TBA

"inter" TBA

"edgelist" as the edgelist for predicted network.

"nedges" as the number of edges of the predicted network.
}
\description{
Choose the optimal regulization parameter and scale paramter for prior information adjusted network construction.
}
\examples{
predictHybNetwork(data = GeneExpresssion, prior = Priorinformation, proteomicResponses = proteomicResponses, nProt = nProt)
}
\concept{zeptosensPkg}
