% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readZeptosensXls.R
\name{readZeptosensXls}
\alias{readZeptosensXls}
\title{Reads the Zeptosens File (.xls) into a Data.Frame}
\usage{
readZeptosensXls(inputFile)
}
\arguments{
\item{inputFile}{}
}
\value{
a data.frame where columns have been classified as "character" or "numeric"
}
\description{
Reads the Zeptosens File (.xls) into a Data.Frame
}
\concept{
zeptosensPkg
}

