% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repAverage.R
\name{repAverage}
\alias{repAverage}
\title{average replicates Zeptosens Data}
\usage{
repAverage(array1)
}
\arguments{
\item{array1}{FIXME}

\item{average_rep}{}

\item{controlProbeIndex}{row number that that contains the first control probe}

\item{antibodyNum}{number of antibodies on array}
}
\description{
average replicates Zeptosens Data
}
\concept{
zeptosensPkg
}

