% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_cross_validation.R
\name{run_cross_validation}
\alias{run_cross_validation}
\title{Run Cross validation for data through the chosen algorithm.}
\usage{
run_cross_validation(data, prior, boot_time = 1000, cut_off = 0.1,
  fold = 5)
}
\arguments{
\item{data}{input expression data. Coloumns as the gene, rows as the sample.With
colnames as the gene tags, rownames as the sample tags.}

\item{prior}{Prior information data frame ,with colnames and rownames as gene tags.
With colnames and rownames as gene tags. Can be inferred from predict_bio_network() or any network resources.}

\item{boot_time}{Bootstrap time mannually set.Default at 1000.}

\item{cut_off}{Manually set up cut off value for strength of edge. (Default at 0.1)}

\item{fold}{The fold for training and test dataset. Default at 5.}
}
\value{
result of validation score. for random network and the network predicted with the algorithm.
}
\description{
Run Cross validation for data through the chosen algorithm.
}
\concept{zeptosensPkg}
