% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCrossValidation.R
\name{runCrossValidation}
\alias{runCrossValidation}
\title{Run Cross validation for data through the chosen algorithm.}
\usage{
runCrossValidation(data, prior, bootTime = 1000, cutOff = 0.1,
  fold = 5)
}
\arguments{
\item{data}{input expression data. Coloumns as the gene, rows as the sample.With
colnames as the gene tags, rownames as the sample tags.}

\item{prior}{prior information matrix with colnames and rownames as the gene tags.}

\item{bootTime}{Bootstrap time mannually set.Default at 1000.}

\item{fold}{The fold for training and test dataset. Default at 5.}
}
\value{
result of validation score. for random network and the network predicted with the algorithm.
}
\description{
Run Cross validation for data through the chosen algorithm.
}
\concept{zeptosensPkg}
