% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samp_sdev.R
\name{samp_sdev}
\alias{samp_sdev}
\title{stdev over all samples}
\usage{
samp_sdev(n_sample, n_prot, n_dose, n_x)
}
\arguments{
\item{n_sample}{samples number of input data.}

\item{n_prot}{Antibody number of input data.}

\item{n_dose}{doses number of input data.}

\item{n_x}{vertically merged data from all samples rows condition columns proteins}
}
\description{
stdev over all samples
}
\concept{zeptosensPkg}
