% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampSdev.R
\name{sampSdev}
\alias{sampSdev}
\title{stdev over all samples}
\usage{
sampSdev(n_sample, n_prot, n_dose, n_x)
}
\arguments{
\item{n_sample}{number of samples}

\item{n_prot}{number of proteins}

\item{n_dose}{number of doses}

\item{n_x}{vertically merged data from all samples rows condition columns proteins}
}
\description{
stdev over all samples
}
\concept{zeptosensPkg}
