% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/totalproteinnormalization.R
\name{totalproteinnormalization}
\alias{totalproteinnormalization}
\title{Normalize Zeptosens Data wrt ttl protein level}
\usage{
totalproteinnormalization(Array1, totalProteinLevels1, nsamples, antibodyNum)
}
\arguments{
\item{totalProteinLevels1}{total protein levels for array 1 (use readZeptosensData format)}

\item{nsamples}{samples/Ab}

\item{antibodyNum}{number of antibodies on array}

\item{array1}{FIXME}
}
\description{
Normalize Zeptosens Data wrt ttl protein level
}
\details{
Normalize wrt ttl prt lvl
}
\concept{
zeptosensPkg
}

