% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSifFromMatrix.R
\name{createSifFromMatrix}
\alias{createSifFromMatrix}
\title{Create sif from network matrix.}
\usage{
createSifFromMatrix(t.net, genelist)
}
\arguments{
\item{t.net}{p*p Gaussian gene Network matrix estimated.}

\item{genelist}{gene name list corresponding to the Gaussian gene Network estimated.}
}
\value{
edgelist decription dataframe contain G(V,E) including Vertex and Edges(1 positive/-1 negative) edgevalue and vertex number of Gaussian Graphical Model.
}
\description{
Create sif from network matrix.
}
\concept{
zeptosensPkg
}
