% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSifFromMatrix.R
\name{createSifFromMatrix}
\alias{createSifFromMatrix}
\title{Create sif from network matrix.}
\usage{
createSifFromMatrix(t.net)
}
\arguments{
\item{t.net}{p*p gene correlation network matrix #}
}
\value{
edgelist decription dataframe contain upperstream gene to downstream gene.#
}
\description{
Create sif from network matrix.
}
\concept{
zeptosensPkg
}
