% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_target_score.R
\name{get_target_score}
\alias{get_target_score}
\title{Target Score Pilot Run}
\usage{
get_target_score(wk, wks, dist_ind, inter, n_dose, n_prot,
  proteomic_responses, max_dist = 1, n_perm, cell_line,
  target_score_output_file = NULL, matrix_wk_output_file = NULL,
  target_score_q_value_file = NULL, target_score_dose_file = NULL,
  random_target_score_file = NULL, target_score_p_value_file = NULL,
  verbose = TRUE, ts_factor = 1, fs_file,
  signed_matrix_wk_output_file = NULL)
}
\arguments{
\item{wk}{TODO}

\item{wks}{TODO}

\item{dist_ind}{TODO}

\item{inter}{TODO}

\item{n_dose}{TODO}

\item{n_prot}{TODO}

\item{proteomic_responses}{TODO}

\item{max_dist}{TODO (default: 1)}

\item{n_perm}{number of random TS calculations for building the null distribution}

\item{cell_line}{TODO}

\item{target_score_output_file}{a filename to write total target score results (default: NULL)}

\item{matrix_wk_output_file}{TODO}

\item{target_score_q_value_file}{a filename to write statistical significance levels (default: NULL)}

\item{target_score_dose_file}{a filename to write dose dependent target score results (default: NULL)}

\item{random_target_score_file}{TODO}

\item{target_score_p_value_file}{a filename to write statistical significance levels (default: NULL)}

\item{verbose}{a flag for debugging output}

\item{ts_factor}{a scaling factor for the pathway component of the target score}

\item{fs_file}{a file with the functional score data}

\item{signed_matrix_wk_output_file}{TODO}
}
\description{
Target Score Pilot Run
}
\details{
data: multiple dose single drug perturbation
ts: integral_dose(fs*(xi+sigma_j(2^p*xj*product_k(wk))))
to be converted to integral_dose(fs*((xi/(stdev_xi)+sigma_j(2^p*((xj/stdev_xj)*product_k(wk))))
missing: For phosp and dephosp based wk, there is no 'exact match' between known and measured phospho-sites
}
\concept{zeptosensPkg}
