% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_target_score.R
\name{get_target_score}
\alias{get_target_score}
\title{Target Score Pilot Run}
\usage{
get_target_score(
  wk,
  wks,
  dist_ind,
  inter,
  n_dose,
  n_prot,
  proteomic_responses,
  n_perm,
  verbose = TRUE,
  ts_factor = 1,
  fs_dat
)
}
\arguments{
\item{wk}{inference network constructed in matrix form with edge strength value estimated.
Can be extracted directly from predict_bio_network,predict_dat_network or predictt_hyb_network
function. Where predict_bio_network edge value default at 1 for upregulate and -1 for down regulate.}

\item{wks}{inference network constructed in matrix form with edge strength value estimated.
Can be extracted directly from predict_bio_network,predict_dat_network or predictt_hyb_network
function. Where predict_bio_network edge value default at 1 for upregulate, -1 for down regulate,
2 for phosphorylation and -2 for dephosphorylation.}

\item{dist_ind}{i distance file of edgelist with a third column as the network distance
between the genes in the interaction}

\item{inter}{edgelist of inferred network.}

\item{n_dose}{dose number of input data.}

\item{n_prot}{antibody number of input data.}

\item{proteomic_responses}{input drug perturbation data. With columns as antibody, rows as samples.}

\item{n_perm}{number of random TS calculations for building the null distribution}

\item{verbose}{a flag for debugging output}

\item{ts_factor}{a scaling factor for the pathway component of the target score}

\item{fs_dat}{a file with the functional score data}
}
\description{
Target Score Pilot Run
}
\details{
data: multiple dose single drug perturbation
ts: integral_dose(fs*(xi+sigma_j(2^p*xj*product_k(wk))))
to be converted to integral_dose(fs*((xi/(stdev_xi)+sigma_j(2^p*((xj/stdev_xj)*product_k(wk))))
missing: For phosp and dephosp based wk, there is no 'exact match' between known and measured phospho-sites
}
\concept{zeptosensPkg}
