% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_volcano_plot.R
\name{get_volcano_plot}
\alias{get_volcano_plot}
\title{Plot volcano plot of Target Score Result.
As Significant Proteins (p value< Defalut or Manually set value) will show in red.
The inverse log 10 of Target Score q value and Target Score calculated were shown.}
\usage{
get_volcano_plot(ts, q_value, filename, path = NULL, sig_value = 0.4)
}
\arguments{
\item{ts}{input Target Score calculated for each protein data frame.
Gene in coloumns and samples in row. With colnames as gene tags and rownames as sample tags.}

\item{q_value}{input Target Score q value calculated for each protein data frame.
Gene in coloumns and samples in row. With colnames as gene tags and rownames as sample tags.}

\item{filename}{Manually set filename of volcano plot.}

\item{path}{Plot Store path. Default at working environment.}

\item{sig_value}{Manually set significant cut-off value for log10(q_value). (Default at 0.4)}
}
\value{
volcano Plots with indicated filename and path.
}
\description{
Plot volcano plot of Target Score Result.
As Significant Proteins (p value< Defalut or Manually set value) will show in red.
The inverse log 10 of Target Score q value and Target Score calculated were shown.
}
\concept{zeptosensPkg}
