% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVolcanoPlot.R
\name{getVolcanoPlot}
\alias{getVolcanoPlot}
\title{Plot volcano plot of Target Score Result.
As Significant Proteins (p value< Defalut or Manually set value) will show in red.
The inverse log 10 of Target Score q value and Target Score calculated were shown.}
\usage{
getVolcanoPlot(ts, qValue, filename, path = NULL, sigValue = 0.4)
}
\arguments{
\item{ts}{input Target Score calculated for each protein data frame.
Gene in coloumns and samples in row. Wih colnames as gene tags and rownames as sample tags.}

\item{qValue}{input Target Score q value calculated for each protein data frame.
Gene in coloumns and samples in row. Wih colnames as gene tags and rownames as sample tags.}

\item{filename}{Manually set filename of volcano Plot.}

\item{path}{Plot Store path. Default at working environment.}
}
\value{
volcano Plots with indicated filename and path.
}
\description{
Plot volcano plot of Target Score Result.
As Significant Proteins (p value< Defalut or Manually set value) will show in red.
The inverse log 10 of Target Score q value and Target Score calculated were shown.
}
\examples{
optimizeParameter(data = GeneExpresssion, prior = Priorindormation)
}
\concept{zeptosensPkg}
