% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network2.R
\name{network2}
\alias{network2}
\title{Network for adjusted-glasso}
\usage{
network2(wk, n_prot, proteomic_responses, dist_file = NULL, verbose = FALSE)
}
\arguments{
\item{wk}{Inference network constructed in matrix form with edge strength value estimated.
Can be extracted directly from predict_bio_network,predict_dat_network or predictt_hyb_network
function. Where predict_bio_network edge value default at 1 for upregulate and -1 for down regulate.}

\item{n_prot}{Antibody number of input data.}

\item{proteomic_responses}{Input drug perturbation data. With columns as antibody, rows as samples.}

\item{dist_file}{A distance file an edgelist with a third column which is the network distance
between the genes in the interaction}

\item{verbose}{logical, whether to show additional debugging information}
}
\description{
Network for adjusted-glasso
}
\concept{zeptosensPkg}
