% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network2.R
\name{network2}
\alias{network2}
\title{Network for adjusted-glasso}
\usage{
network2(wk, n_prot, proteomic_responses, max_dist,
  antibody_map_file = NULL, dist_file = NULL, verbose = FALSE)
}
\arguments{
\item{wk}{The upload network constructed In matrix form.}

\item{n_prot}{number of proteins tested.}

\item{proteomic_responses}{TODO}

\item{max_dist}{TODO}

\item{antibody_map_file}{TODO}

\item{dist_file}{A distance file an edgelist with a third column which is the network distance
between the genes in the interaction}

\item{verbose}{logical, whether to show additional debugging information}
}
\description{
Network for adjusted-glasso
}
\concept{zeptosensPkg}
