% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network2.R
\name{network2}
\alias{network2}
\title{Network for adjusted-glasso}
\usage{
network2(wk, nProt, proteomicResponses, maxDist, antibodyMapFile = NULL,
  distFile = NULL, verbose = F)
}
\arguments{
\item{wk}{The upload network constructed In matrix form.}

\item{nProt}{number of proteins tested.}

\item{proteomicResponses}{TBA}

\item{distFile}{A distance file an edgelist with a third column which is the network distance
between the genes in the interaction}
}
\description{
Network for adjusted-glasso
}
\examples{

}
\concept{zeptosensPkg}
