% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network2.R
\name{network2}
\alias{network2}
\title{Network for adjusted-glasso}
\usage{
network2(wk, n_prot, proteomic_responses, dist_file = NULL, verbose = FALSE)
}
\arguments{
\item{wk}{Inference network constructed in matrix form with edge strength value estimated.
Can be extracted directly from predict_bio_network,predict_dat_network or predictt_hyb_network
function. Where predict_bio_network edge value default at 1 for upregulate and -1 for down regulate.}

\item{n_prot}{Antibody number of input data.}

\item{proteomic_responses}{Input drug perturbation data. With columns as antibody, rows as samples.}

\item{dist_file}{A distance file an edgelist with a third column which is the network distance
between the genes in the interaction}

\item{verbose}{logical, whether to show additional debugging information}
}
\value{
a list is returned with the following entries:
* "wk" inferred network matrix form with edge strength value default at 1 for
upregulate and -1 for down regulate.
* "wks" as inferred network matrix form with edge strength value default at 1 for
upregulate and -1 for down regulate and 2 for phosphorylation and -2 for dephosphorylation.
* "dist_ind" A distance file of edgelist with a third column as the network distance
between the genes in the interaction.
* "inter" file as edgelist of inferred network.
}
\description{
Network for adjusted-glasso
}
\examples{
# Read proteomic response for cellline1
file <- system.file("test_data", "BT474.csv", package = "zeptosensPkg")
proteomic_responses <- read.csv(file, row.names = 1)
 
 # Read in network output
 wk_org <- readRDS(system.file("test_data_files", "predict_hybrid_network_network_output.rds",
 package = "zeptosensPkg"
 ))
 
 network <- zeptosensPkg::network2(
 wk <- wk_org$wk,
 n_prot = dim(proteomic_responses)[2],
 proteomic_responses = proteomic_responses
 )

}
\concept{zeptosensPkg}
