% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_parameter_dat.R
\name{optimize_parameter_dat}
\alias{optimize_parameter_dat}
\title{Choose the optimal regulization parameter for only data-driven network.}
\usage{
optimize_parameter_dat(data, rho = 10^seq(-2, 0, 0.02))
}
\arguments{
\item{data}{input proteomics dataset for network inference. Gene in coloumns and samples in row.
With colnames as gene tags and rownames as sample tags.}

\item{rho}{positive tuning parameter for elastic net penalty. Default at 10^seq(-2, 0, 0.02).}
}
\value{
Parameter of regulization BIC error calculated.
\item{rho}{penalty parameter optimized.}
\item{bic}{BIC error calculated for penalty parameters.}
}
\description{
Choose the optimal regulization parameter for only data-driven network.
}
\examples{
# read proteomic responce file
signaling_responses <- read.csv(system.file("test_data", "TCGA-BRCA-L4.csv", package = "zeptosensPkg"), row.names = 1)
 parameters <- zeptosensPkg::optimize_parameter_dat(data = signaling_responses)

}
\concept{zeptosensPkg}
