% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_parameter_dat.R
\name{optimize_parameter_dat}
\alias{optimize_parameter_dat}
\title{Choose the optimal regulization parameter for only data-driven network.}
\usage{
optimize_parameter_dat(data)
}
\arguments{
\item{data}{input proteomics dataset for network inference. Gene in coloumns and samples in row.
With colnames as gene tags and rownames as sample tags.}
}
\value{
Parameter of regulization decided lowest BIC. Including regularize parameter
(L1 norm parameter) as rho.
}
\description{
Choose the optimal regulization parameter for only data-driven network.
}
\concept{zeptosensPkg}
