% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizeParameter.R
\name{optimizeParameter}
\alias{optimizeParameter}
\title{Choose the optimal regulization parameter and scale paramter.}
\usage{
optimizeParameter(data, prior)
}
\arguments{
\item{data}{input expression data. Gene in coloumns and samples in row. Wih colnames as gene tags and rownames as sample tags.}

\item{prior}{Prior information matrix,with colnames and rownames as gene tags.}
}
\value{
Parameter list of regulization parameter decided by the prior information and the algorithmn lowest BIC.Including regularize parameter as rho, scale parameter as kappa, and regulization matrix decided.
}
\description{
Choose the optimal regulization parameter and scale paramter.
}
\concept{
zeptosensPkg
}
