% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_bio_network.R
\name{predict_bio_network}
\alias{predict_bio_network}
\title{Predict Network from signedPC (Bio-inferred network)}
\usage{
predict_bio_network(n_prot, proteomic_responses, max_dist,
  antibody_map_file = NULL, dist_file = NULL, verbose = FALSE)
}
\arguments{
\item{n_prot}{TODO}

\item{proteomic_responses}{TODO}

\item{max_dist}{TODO (default: 1)}

\item{antibody_map_file}{a listing of antibodies, their associated genes, and modification sites}

\item{dist_file}{A distance file an edgelist with a third column which is the network distance
between the genes in the interaction}

\item{verbose}{whether to show debugging information}
}
\value{
TODO
}
\description{
Predict Network from signedPC (Bio-inferred network)
}
\concept{zeptosensPkg}
