% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_bio_network.R
\name{predict_bio_network}
\alias{predict_bio_network}
\title{Predict Network from signedPC (Bio-inferred network)}
\usage{
predict_bio_network(
  n_prot,
  proteomic_responses,
  max_dist,
  mab_to_genes,
  dist_file = NULL,
  verbose = FALSE
)
}
\arguments{
\item{n_prot}{Antibody number of input data.}

\item{proteomic_responses}{Input drug perturbation data. With columns as antibody, rows as samples.}

\item{max_dist}{maximum distance between two antibody. (Default at 1)}

\item{mab_to_genes}{A list of antibodies, their associated genes, modification sites and effect.}

\item{dist_file}{A distance file an edgelist with a third column which is the network distance
between the genes in the interaction}

\item{verbose}{whether to show debugging information}
}
\value{
a list is returned with the following entries:
* "wk" inferred network matrix form with edge strength value default at 1 for
upregulate and -1 for down regulate.
* "wks" as inferred network matrix form with edge strength value default at 1 for
upregulate and -1 for down regulate and 2 for phosphorylation and -2 for dephosphorylation.
* "dist_ind" A distance file of edgelist with a third column as the network distance
between the genes in the interaction.
* "inter" file as edgelist of inferred network.
* "edgelist" file as sif file of edgelist for inferred network.
}
\description{
Predict Network from signedPC (Bio-inferred network)
}
\examples{
# READ ANTIBODY FILE ----
mab_to_genes <- read.table(system.file("targetscoreData", "antibodyMapFile_08092019.txt", package = "zeptosensPkg"),
sep = "\t",
header = TRUE,
stringsAsFactors = FALSE
)

# Read proteomic response for cellline1
proteomic_responses <- read.csv(system.file("test_data", "BT474.csv", package = "zeptosensPkg"), row.names = 1)

 # Extract network
 network <-  predict_bio_network(
 n_prot = dim(proteomic_responses)[2],
 proteomic_responses = proteomic_responses,
 max_dist = 1,
 mab_to_genes = mab_to_genes
 )

}
\concept{zeptosensPkg}
