% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictDatNetwork.R
\name{predictDatNetwork}
\alias{predictDatNetwork}
\title{predict data-driven only network.}
\usage{
predictDatNetwork(data, cut_off = 0.1, nProt, proteomicResponses,
  maxDist = 1)
}
\arguments{
\item{data}{input expression data frame. Gene in coloumns and samples in row. Wih colnames as gene tags and rownames as sample tags.}

\item{cut_off}{Manually Setup cut off value for the strength of edge. Default at 0.1.}

\item{nProt}{Number of proteins included in the drug perturbation data.}

\item{proteomicResponses}{Drug Perturbation data for analysis.}

\item{maxDist}{TBA (default: 1)}
}
\value{
Parameter of regulization decided lowest BIC.Including regularize parameter(L1 norm parameter) as rho.
}
\description{
predict data-driven only network.
}
\examples{
optimizeParameter(data=GeneExpresssion,prior=Priorindormation)
}
\concept{zeptosensPkg}
