% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_cross_validation.R
\name{run_cross_validation}
\alias{run_cross_validation}
\title{Run cross validation through data.}
\usage{
run_cross_validation(data, prior = NULL, n_prot, proteomic_responses,
  mab_to_genes, dist_file = NULL, max_dist = 1, verbose = FALSE,
  boot_time, k_fold = 5, crit_cv)
}
\arguments{
\item{data}{input proteomics dataset for network inference. Gene in coloumns and samples in row.
With colnames as gene tags and rownames as sample tags.}

\item{prior}{Prior information matrix of gene interaction, with colnames and rownames as gene tags.
With colnames and rownames as gene tags.Can be inferred from Public data source (for example:SignedPC).}

\item{n_prot}{Antibody number of input data.}

\item{proteomic_responses}{Input drug perturbation data. With columns as antibody, rows as samples.}

\item{mab_to_genes}{A list of antibodies, their associated genes, modification sites and effect.}

\item{dist_file}{A distance file an edgelist with a third column which is the network distance
between the genes in the interaction}

\item{max_dist}{maximum distance between two antibody. (Default at 1)}

\item{verbose}{whether to show debugging information}

\item{boot_time}{Bootstrappin number. (Default at 1000)}
}
\value{
Parameter list of regulization parameter decided by the prior information and the algorithmn lowest BIC.
\item{cv_error_dat}
\item{cv_error_hyb}
\item{diff_cv_error}
\item{mean_cv_error_dat}
\item{mean_cv_error_hyb}
\item{mean_cv_error_dif}
\item{sd_cv_error_dat}
\item{sd_cv_error_hyb}
\item{sd_cv_error_dif}
}
\description{
Run cross validation through data.
}
\concept{zeptosensPkg}
