% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_cross_validation.R
\name{run_cross_validation}
\alias{run_cross_validation}
\title{Run cross validation through data.}
\usage{
run_cross_validation(
  data,
  prior = NULL,
  n_prot,
  k_fold = 5,
  proteomic_responses,
  mab_to_genes,
  dist_file = NULL,
  max_dist = 1,
  verbose = FALSE,
  boot_time = 100,
  crit_cv = "loglik",
  algorithm = c("data_driven", "hybrid_driven")
)
}
\arguments{
\item{data}{input proteomics dataset for network inference. Gene in coloumns and samples in row.
With colnames as gene tags and rownames as sample tags.}

\item{prior}{prior information matrix of gene interaction, with colnames and rownames as gene tags.
With colnames and rownames as gene tags.Can be inferred from Public data source (for example:SignedPC).}

\item{n_prot}{antibody number of input data.}

\item{k_fold}{specify the number of folds for cross validation.(Default at 5)}

\item{proteomic_responses}{input drug perturbation data. With columns as antibody, rows as samples.}

\item{mab_to_genes}{A list of antibodies, their associated genes, modification sites and effect.}

\item{dist_file}{a distance file an edgelist with a third column which is the network distance
between the genes in the interaction}

\item{max_dist}{maximum distance between two antibody. (Default at 1)}

\item{verbose}{whether to show debugging information}

\item{boot_time}{bootstrap time mannually set.Default at 100.}

\item{crit_cv}{cross validation criterion (\code{loglik}, \code{AIC}, or \code{BIC}). Defaults to \code{loglik}.}

\item{algorithm}{flexible toolbox implementing network estimating algorithms for robustness test.
(\code{data_driven},or \code{hybrid_driven}).}
}
\value{
returns list of cross validation results which includes:
\item{cv_error}{cross validation errors of data.}
\item{cv_error_r}{cross validation errors of random data.}
\item{cv_error_diff}{cross validation errors difference between random data and data.}
}
\description{
Run cross validation through data.
}
\concept{zeptosensPkg}
