% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCrossValidation.R
\name{runCrossValidation}
\alias{runCrossValidation}
\title{Run 10-fold Cross validation for data through the current algorythm.}
\usage{
runCrossValidation(data, prior, boot.time, rho, kappa, cut.off)
}
\arguments{
\item{data}{input expression data. Coloumns as the gene, rows as the sample.With colnames as the gene tags, rownames as the sample tags.}

\item{prior}{prior information matrix with colnames and rownames as the gene tags.}

\item{boot.time}{Bootstrap time mannually set.}

\item{rho}{regulization parameter}

\item{kappa}{scaler parameter}
}
\value{
result lists of positive predictive values, negative predictive values,sensitivity, specificity,for random network and the network predicted with the algorithm.
}
\description{
Run 10-fold Cross validation for data through the current algorythm.
}
\concept{
zeptosensPkg
}
