% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samp_sdev.R
\name{samp_sdev}
\alias{samp_sdev}
\title{Calculate the standard deviation over all samples}
\usage{
samp_sdev(n_sample, n_prot, n_dose, n_x, replace_missing = TRUE)
}
\arguments{
\item{n_sample}{samples number of input data.}

\item{n_prot}{Antibody number of input data.}

\item{n_dose}{doses number of input data.}

\item{n_x}{vertically merged data from all samples rows condition columns proteins}

\item{replace_missing}{Replace missing values NA with 0 (DEFAULT: TRUE)}
}
\description{
Calculate the standard deviation over all samples
}
\examples{
# read proteomic responce file
file <- system.file("test_data", "BT474.csv", package = "zeptosensPkg")
proteomic_responses <- read.csv(file, row.names = 1)
samp_d <- samp_sdev(
n_x = proteomic_responses, n_sample = dim(proteomic_responses)[1],
n_prot = dim(proteomic_responses)[2], n_dose = 1
)

}
\concept{zeptosensPkg}
