% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampSdev.R
\name{sampSdev}
\alias{sampSdev}
\title{stdev over all samples}
\usage{
sampSdev(nSample, nProt, nDose, nX)
}
\arguments{
\item{nSample}{number of samples}

\item{nProt}{number of proteins}

\item{nX}{vertically merged data from all samples rows condition columns proteins}

\item{nCond}{number of consitions}
}
\description{
stdev over all samples
}
\concept{
zeptosensPkg
}
